Add N64 missing instructions test into CI and fix issues. Add test to make sure naken_asm / naken_util doesn't require libstc++ to run.

+++ tests/other/n64_rsp_illegal_instr.c
@@ -128,7 +128,7 @@ int main(int argc, char *argv[])
 
   int n;
 
-  for (n = 0; n < sizeof(instructions) / sizeof(const char *); n++)
+  for (n = 0; n < (int)(sizeof(instructions) / sizeof(const char *)); n++)
   {
     errors += test_instruction(instructions[n]);
   }
